Move dockerstats struct
